+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Remember to align the itemized text with the first line of an item within a list
   * Breaking changes
     * Deleted {func}`jax.scipy.linalg.polar_unitary`, which was a deprecated JAX 
       extension to the scipy API. Use {func}`jax.scipy.linalg.polar` instead.
+    * In presence of jax2tf shape polymorphism, non-constant dimension sizes are
+      not hashable anymore ({jax-issue}`#14200`).
 
 ## jaxlib 0.4.3